refactor mx quantizer
