+++ gemlite/helper.py
@@ -1,12 +1,12 @@
 # Written by Dr. Hicham Badri @Mobius Labs GmbH - 2024
 #********************************************************
-
 import torch
 import gc
 import time
 from tqdm import tqdm
 from gemlite.core import GemLiteLinearTriton, DType, TORCH_TO_DTYPE
 from gemlite.triton_kernels.utils import M_MAPPING
+from gemlite.quant_utils import WeightQuantizerMXFP
 from .triton_kernels.utils import IS_HIP
 
 if IS_HIP:
@@ -26,123 +26,6 @@ def cleanup_linear(linear_layer, del_orig=True):
             if(hasattr(linear_layer, attr)):
                 setattr(linear_layer, attr, None)
     torch.cuda.empty_cache()
-
-class QuantizerMXFP:
-    def __init__(self, compute_dtype=torch.bfloat16, device="cuda:0"):
-        self.compute_dtype  = compute_dtype
-        self.device         = device
-        self.fp4_values     = torch.tensor([0, 0.5, 1, 1.5, 2, 3, 4, 6, -0, -0.5, -1, -1.5, -2, -3, -4, -6], dtype=compute_dtype, device=device) 
-        self.fp4_p_vals     = torch.tensor([0, 0.5, 1, 1.5, 2, 3, 4, 6], dtype=compute_dtype, device=device)
-        self.fp4_thresholds = torch.tensor([0.25, 0.75, 1.25, 1.75, 2.5, 3.5, 5.], dtype=compute_dtype, device=device)  #(fp4_p_vals[:-1] + fp4_p_vals[1:]) / 2
-        self.max_val        = self.fp4_values.abs().max()
-
-    def round_to_closest_fp4(self, tensor):
-        out = self.fp4_p_vals[torch.searchsorted(self.fp4_thresholds, tensor.abs())]
-        out *= tensor.sign()
-        return out
-
-    def to_index(self, W_q):
-        assert W_q.is_floating_point(), "Input should be floating point fp4 values."
-        return (W_q.view(-1, 1) == self.fp4_values.view(1 , -1)).to(torch.uint8).argmax(dim=1).to(torch.uint8).view(W_q.shape)
-
-    @torch.compile(fullgraph=True)
-    def quantize_mxfp8(self, W: torch.Tensor, eps: float = 2**-30, index: bool = False, mx_fp8_dtype: torch.dtype = torch.float8_e4m3fn) -> (torch.Tensor, torch.Tensor):
-        group_size: int = 32
-        max_val = torch.finfo(mx_fp8_dtype).max
-
-        W_flat = W.view(-1, group_size).float()
-        ideal_scale = W_flat.abs().amax(dim=1, keepdim=True)
-        ideal_scale /= max_val
-
-        scales = (2 ** torch.ceil(torch.log2(ideal_scale))).clamp_(min=eps)
-
-        W_q = (W_flat / scales).to(mx_fp8_dtype).to(W_flat.dtype)
-        scales = scales.to(torch.float8_e8m0fnu)
-
-        if(index):
-            W_q = W_q.to(mx_fp8_dtype)
-
-        return W_q, scales
-
-    @torch.compile(fullgraph=True)
-    def quantize_mxfp4(self, W: torch.Tensor, window_size: int = 0, eps: float = 2**-30, index: bool = False) -> (torch.Tensor, torch.Tensor):
-        group_size: int = 32
-        W_nbits = 4
-        max_val = 6
-
-        W_flat = W.view(-1, group_size).float()
-        ideal_scale = W_flat.abs().amax(dim=1, keepdim=True)
-        ideal_scale /= max_val
-
-        if(window_size == 0):
-            scales = 2 ** torch.ceil(torch.log2(ideal_scale))
-        else:
-            initial_log2_scales = torch.ceil(torch.log2(ideal_scale))
-            search_offsets = torch.arange(-window_size, window_size + 1, device=W.device, dtype=initial_log2_scales.dtype).view(1, -1)
-            candidate_scales = torch.pow(2, initial_log2_scales + search_offsets)
-            candidate_scales[candidate_scales < eps] = eps
-
-            W_q_candidates = self.round_to_closest_fp4(W_flat.unsqueeze(1) / candidate_scales.unsqueeze(-1))
-            W_r_candidates = W_q_candidates * candidate_scales.unsqueeze(-1)
-            errors = (W_flat.unsqueeze(1) - W_r_candidates).abs().mean(dim=-1)
-            scales = torch.gather(candidate_scales, 1, torch.argmin(errors, dim=1, keepdim=True))
-
-        scales = scales.clamp_(eps)
-        W_q = self.round_to_closest_fp4(W_flat / scales)
-        scales = scales.to(torch.float8_e8m0fnu)
-
-        if(index):
-            W_q = self.to_index(W_q)
-        return W_q, scales
-
-    @torch.compile(fullgraph=True)
-    def quantize_nvfp4(self, W: torch.Tensor, window_size: int = 0, eps: float = 1e-8, index: bool = False) -> (torch.Tensor, torch.Tensor):
-        group_size: int = 16
-        W_nbits = 4
-        max_val = 6
-        fp8_dtype = torch.float8_e4m3fn
-        max_fp8 = torch.finfo(fp8_dtype).max #448
-
-        W_flat = W.view(-1, group_size).float()
-        ideal_scale = W_flat.abs().amax(dim=1, keepdim=True)
-        ideal_scale /= max_val
-        ideal_scale.clamp_(max=max_fp8)
-
-        if(window_size == 0):
-            scales = ideal_scale.to(fp8_dtype).to(ideal_scale.dtype)
-        else:
-            search_offsets = torch.arange(-window_size, window_size + 1, device=W.device, dtype=torch.int).view(1, -1)
-            candidate_scales = (ideal_scale.to(fp8_dtype).view(torch.int8) + search_offsets).clamp_(-128, 127).to(torch.int8)
-
-            #Avoid nan in int8 range (-1, 127 as int8 as e4m3 nans)
-            candidate_scales[candidate_scales==-1] = 1
-            candidate_scales[candidate_scales==127] = 1
-            candidate_scales = candidate_scales.view(fp8_dtype).float()
-            candidate_scales[candidate_scales < eps] = eps
-
-            W_q_candidates = self.round_to_closest_fp4(W_flat.unsqueeze(1) / candidate_scales.unsqueeze(-1))
-            W_r_candidates = W_q_candidates * candidate_scales.unsqueeze(-1)
-            errors = (W_flat.unsqueeze(1) - W_r_candidates).abs().mean(dim=-1)
-            scales = torch.gather(candidate_scales, 1, torch.argmin(errors, dim=1, keepdim=True))
-
-        scales = scales.clamp_(min=eps)
-        W_q = self.round_to_closest_fp4(W_flat / scales)
-        scales = scales.to(fp8_dtype)
-
-        if(index):
-            W_q = self.to_index(W_q)
-
-        return W_q, scales
-
-    def dequantize(self, W_q, scales, shape = None, dtype = None):
-        if(W_q.dtype == torch.uint8): #from indices
-            W_q = self.fp4_values[W_q.int()]
-
-        group_size = W_q.numel() // scales.numel()
-        out = (W_q.view([-1, group_size]).float() * scales.float())
-        if(shape is not None):
-            out = out.view(shape)
-        return out.to(self.compute_dtype if dtype is None else dtype)
 
 #Replaces all linear layers with the corresponding processor
 def patch_model(model, device, processor, skip_modules=[]):
@@ -357,7 +240,7 @@ class A16Wn: #8/4/2-bit weight-only as grouped "INT" / 8/4-bit as MXFP type
 
     def mxfp_from_linear(self, linear_layer, W_nbits, del_orig=True):
         if(self.quantizer_mx is None):
-            self.quantizer_mx = QuantizerMXFP(device=self.device, compute_dtype=linear_layer.weight.dtype)
+            self.quantizer_mx = WeightQuantizerMXFP(device=self.device, compute_dtype=linear_layer.weight.dtype)
 
         W = linear_layer.weight.data
         bias = linear_layer.bias.clone() if (linear_layer.bias is not None) else None 
@@ -566,7 +449,7 @@ class A8W8_MXFP_dynamic:
 
     def from_linear(self, linear_layer, del_orig=True):
         if(self.quantizer_mx is None):
-            self.quantizer_mx = QuantizerMXFP(device=self.device, compute_dtype=linear_layer.weight.dtype)
+            self.quantizer_mx = WeightQuantizerMXFP(device=self.device, compute_dtype=linear_layer.weight.dtype)
 
         W = linear_layer.weight.data
         bias = linear_layer.bias.clone() if (linear_layer.bias is not None) else None 
@@ -736,7 +619,7 @@ class A8Wn_MXFP_dynamic:
 
     def from_linear(self, linear_layer, del_orig=True):
         if(self.quantizer_mx is None):
-            self.quantizer_mx = QuantizerMXFP(device=self.device, compute_dtype=linear_layer.weight.dtype)
+            self.quantizer_mx = WeightQuantizerMXFP(device=self.device, compute_dtype=linear_layer.weight.dtype)
 
         W = linear_layer.weight.data
         bias = linear_layer.bias.clone() if (linear_layer.bias is not None) else None 
@@ -814,7 +697,7 @@ class A4W4_MXFP_dynamic:
 
     def from_linear(self, linear_layer, del_orig=True):
         if(self.quantizer_mx is None):
-            self.quantizer_mx = QuantizerMXFP(device=self.device, compute_dtype=linear_layer.weight.dtype)
+            self.quantizer_mx = WeightQuantizerMXFP(device=self.device, compute_dtype=linear_layer.weight.dtype)
 
         W = linear_layer.weight.data
         bias = linear_layer.bias.clone() if (linear_layer.bias is not None) else None 
@@ -880,7 +763,7 @@ class A4W4_NVFP_dynamic:
 
     def from_linear(self, linear_layer, del_orig=True):
         if(self.quantizer_mx is None):
-            self.quantizer_mx = QuantizerMXFP(device=self.device, compute_dtype=linear_layer.weight.dtype)
+            self.quantizer_mx = WeightQuantizerMXFP(device=self.device, compute_dtype=linear_layer.weight.dtype)
 
         W = linear_layer.weight.data
         bias = linear_layer.bias.clone() if (linear_layer.bias is not None) else None 
@@ -898,7 +781,7 @@ class A4W4_NVFP_dynamic:
 
 ############################################################################################################################################################
 #BitNet
-class A16W158:
+class A16W158_INT:
     def __init__(self, device='cuda:0', dtype=None, fp32_scale=True):
         self.device = device
         self.dtype = dtype
@@ -945,7 +828,7 @@ class A16W158:
         cleanup_linear(linear_layer, del_orig)
         return out_layer
 
-class A8W158_dynamic:
+class A8W158_INT_dynamic:
     def __init__(self, device='cuda:0', dtype=None, fp32_scale=True):
         self.device = device
         self.dtype = dtype
